refactor: split db.ts by responsibility
- token.ts: hashToken(), generateToken() (pure crypto)
- db.ts: Sql type, createDb(), ensureSchema() (postgres infrastructure)
- ws.ts: inline ConnectorRow + lookupByTokenHash (only consumer)
