+++ packages/relay/src/token.ts
@@ -1,0 +1,9 @@
+import { createHash, randomBytes } from 'node:crypto';
+
+export function hashToken(token: string): string {
+  return createHash('sha256').update(token).digest('hex');
+}
+
+export function generateToken(): string {
+  return randomBytes(32).toString('hex');
+}
